Ability to prevent tests from running against old versions.
With the new mixed-mode tests being run as part of PRB, we'll need
a way to say that a test should not run against old versions.
This adds support for disabling an entire yaml file.
Future steps would be to allow disabling a specific query, and also
having the build update the `!current_version` to the version being
released.